requirements.txt
Django==2.1.5
django-cleanup==3.0.1
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
pytz==2018.9
six==1.12.0
psycopg2==2.7.7
